Turned off destructuring errors in lint
Was going in circles attempting to appease the destructuring error and then violating "no use" of a state variable in the App,jsx file. Needed to pass the state of the navbar to children components. When using this.state.navberOpen to pass, the destructuring error occured. If I tried to  destructure by declaring const { items } = this.state; and passing items.navbarOpen then I got the no use of state variable error.